+++ src/main/java/com/github/lopoha/coboltruffle/builtins/CobolBuiltinNode.java
@@ -1,4 +1,4 @@
-package com.github.lopoha.coboltruffle.parser.builtins;
+package com.github.lopoha.coboltruffle.builtins;
 
 import com.github.lopoha.coboltruffle.parser.CobolException;
 import com.github.lopoha.coboltruffle.parser.CobolExpressionNode;
