+++ src/main/java/com/github/lopoha/coboltruffle/nodes/CobolUndefinedSectionRootNode.java
@@ -1,8 +1,8 @@
-package com.github.lopoha.coboltruffle.parser.nodes;
+package com.github.lopoha.coboltruffle.nodes;
 
 import com.github.lopoha.coboltruffle.parser.CobolLanguage;
 import com.github.lopoha.coboltruffle.parser.CobolRootNode;
-import com.github.lopoha.coboltruffle.parser.runtime.CobolUndefinedNameException;
+import com.github.lopoha.coboltruffle.runtime.CobolUndefinedNameException;
 import com.oracle.truffle.api.frame.VirtualFrame;
 
 public class CobolUndefinedSectionRootNode extends CobolRootNode {
