+++ src/main/java/com/github/lopoha/coboltruffle/parser/CobolEvalRootNode.java
@@ -1,6 +1,5 @@
 package com.github.lopoha.coboltruffle.parser;
 
-import com.github.lopoha.coboltruffle.parser.runtime.CobolNull;
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.RootCallTarget;
