+++ src/main/java/com/github/lopoha/coboltruffle/parser/CobolSectionBodyNode.java
@@ -1,6 +1,6 @@
 package com.github.lopoha.coboltruffle.parser;
 
-import com.github.lopoha.coboltruffle.parser.runtime.CobolNull;
+import com.github.lopoha.coboltruffle.runtime.CobolNull;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.NodeInfo;
